Fix transform args usecase
In case when the set helper is used with args the path contains data
different from the path generated by b.path. (Tested with Ember 3.17)
Example:
```
{{set @greeting.hi ...}}
```
original path has
```
parts: [ 'greeting', 'hi' ],
data: true
```
path generated as b.path('@greeting') has:
```
parts: [ '@greeting' ],
data: false
```

So, it appears that Ember 3.17 processes args in a different way.
And it is not documented anywhere as it is details of implementation.

The best solution I can think of here is to use less destructive AST
transform. Instead of creating new path node, just modify the existing
node: pop the last part from the parts and insert it as a StringLiteral
node and modify path.original to cut off `.key`.